requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
pytest>=7.3.1
requests>=2.28.2
loguru>=0.7.0
tenacity>=8.2.3 